requirements.txt
redis[hiredis]
pika
uvicorn[standard]
fastapi
gensim
invisible_watermark 
transformers 
accelerate 
safetensors
diffusers
pillow
nltk